MiniBank/requirements.txt
# --- Podstawa API ---
fastapi
uvicorn[standard]
pydantic
requests

# --- Baza danych (Async) ---
sqlalchemy
asyncpg

# --- Bezpieczeństwo i JWT ---
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# --- Testowanie (z Labów 6/7) ---
pytest
pytest-asyncio
httpx
testcontainers